Restructure wego-server files and improve ui proxy
