+++ cmd/wego-server/main.go
@@ -1,12 +1,18 @@
 package main
 
 import (
+	"context"
+	"fmt"
 	"math/rand"
+	"net/http"
 	"os"
 	"time"
 
+	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	log "github.com/sirupsen/logrus"
-	"github.com/weaveworks/weave-gitops/cmd/wego-server/app"
+	"github.com/spf13/cobra"
+	pb "github.com/weaveworks/weave-gitops/pkg/api/applications"
+	"github.com/weaveworks/weave-gitops/pkg/server"
 )
 
 func init() {
@@ -22,9 +28,56 @@ func init() {
 func main() {
 	rand.Seed(time.Now().UnixNano())
 
-	command := app.NewAPIServerCommand()
+	command := NewAPIServerCommand()
 
 	if err := command.Execute(); err != nil {
 		os.Exit(1)
 	}
+}
+
+func NewAPIServerCommand() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:  "wego-server",
+		Long: `The wego-server handles HTTP requests for Weave GitOps Applications`,
+
+		SilenceUsage: true,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return StartServer()
+		},
+	}
+	return cmd
+}
+
+var addr = "0.0.0.0:8000"
+
+func StartServer() error {
+	log.Infof("wego api server started. listen address: %s", addr)
+	return RunInProcessGateway(context.Background(), addr)
+}
+
+// RunInProcessGateway starts the invoke in process http gateway.
+func RunInProcessGateway(ctx context.Context, addr string, opts ...runtime.ServeMuxOption) error {
+	mux := runtime.NewServeMux(opts...)
+
+	if err := pb.RegisterApplicationsHandlerServer(ctx, mux, server.NewApplicationsServer()); err != nil {
+		return fmt.Errorf("could not register application: %w", err)
+	}
+	s := &http.Server{
+		Addr:    addr,
+		Handler: mux,
+	}
+
+	go func() {
+		<-ctx.Done()
+		log.Infof("Shutting down the http gateway server")
+		if err := s.Shutdown(context.Background()); err != nil {
+			log.Errorf("Failed to shutdown http gateway server: %v", err)
+		}
+	}()
+
+	if err := s.ListenAndServe(); err != http.ErrServerClosed {
+		log.Errorf("Failed to listen and serve: %v", err)
+		return err
+	}
+	return nil
 }
